# go-backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Tanzanite E-commerce API
  description: |
    完整的电商后端 API 系统，从 WordPress 迁移到 Go。
    
    ## 功能特性
    - 用户认证和授权 (JWT)
    - 产品和内容管理
    - 购物车和订单系统
    - 支付和物流管理
    - 营销系统 (优惠券、积分、会员)
    - 评价和客服系统
    - 产品注册和保修
    - 审计日志
    
    ## 认证
    大部分 API 需要 JWT Token 认证。在请求头中添加：
    ```
    Authorization: Bearer <your_jwt_token>
    ```
    
    ## 速率限制
    - 默认: 100 请求/分钟
    - 管理员: 200 请求/分钟
    
  version: 1.3.0
  contact:
    name: API Support
    email: support@tanzanite.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: 开发环境
  - url: https://api.tanzanite.com
    description: 生产环境
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Token 认证
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: 错误信息
        code:
          type: string
          description: 错误代码
      required:
        - error
...
paths:
...
  /api/v1/marketing/coupons/validate:
    post:
      tags:
        - Marketing
      summary: 验证优惠券
      description: 验证优惠券是否可用
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
                - order_amount
              properties:
                code:
                  type: string
                  example: SUMMER2024
                order_amount:
                  type: number
                  format: float
                  example: 100.00
      responses:
        '200':
          description: 优惠券有效
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                  discount:
                    type: number
                    format: float
                  coupon:
                    type: object
        '400':
          description: 优惠券无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
security:
  - BearerAuth: []